requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.2

